add D457 support for new FW version and fix PR comments

+++ src/platform/hid-data.h
@@ -12,11 +12,8 @@ namespace librealsense {
 struct hid_data
 {
     int32_t x;
-    //uint16_t reserved1[2];
     int32_t y;
-    //uint16_t reserved2[2];
     int32_t z;
-    //uint16_t reserved3[2];
 };
 
 
@@ -25,9 +22,9 @@ struct hid_mipi_data
     uint8_t typeID;
     uint8_t skip1;
     uint64_t hwTs;
-    int16_t x;
-    int16_t y;
-    int16_t z;
+    int32_t x;
+    int32_t y;
+    int32_t z;
     uint64_t hwTs2;
     uint64_t skip2;
 };
